refactor: 예외 핸들링 추가
