+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -1,8 +1,8 @@
 package com.techcourse.controller;
 
-import com.techcourse.Pages;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.Pages;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.web.annotation.Controller;
