+++ app/src/main/java/com/techcourse/controller/LogInOutController.java
@@ -1,11 +1,11 @@
 package com.techcourse.controller;
 
-import com.techcourse.Pages;
 import com.techcourse.domain.UserSession;
 import com.techcourse.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
+import nextstep.mvc.Pages;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.web.annotation.Autowired;
