+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,11 +1,11 @@
 package com.techcourse.controller;
 
-import com.techcourse.Pages;
 import com.techcourse.domain.User;
 import com.techcourse.domain.UserSession;
 import com.techcourse.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.Pages;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.web.annotation.Autowired;
