+++ docs/README.md
@@ -12,7 +12,8 @@
 
 
 ## 해야할것
-- [ ] 커스텀 예외처리 추가하기
-- [ ] 테스트 코드 추가하기
-- [ ] 서비스 컨트롤러 분리하기
+- [x] 커스텀 예외처리 추가하기
+- [] 테스트 코드 추가하기
+- [x] 서비스 컨트롤러 분리하기
 - [ ] Filter 공부하기
+- [x] 필드 DI 구현하기
