+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,8 +4,10 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.controller.exception.AbstractCustomException;
 import nextstep.mvc.controller.tobe.HandlerAdapterRegistry;
 import nextstep.mvc.controller.tobe.HandlerMappingRegistry;
+import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.mvc.view.View;
 import org.slf4j.Logger;
@@ -33,17 +35,25 @@ public class DispatcherServlet extends HttpServlet {
     protected void service(HttpServletRequest request, HttpServletResponse response)
         throws ServletException {
         LOG.debug("Method : {}, Request URI : {}", request.getMethod(), request.getRequestURI());
-
+        ModelAndView modelAndView;
         try {
-            Object handle = handlerMappingRegistry.getHandle(request);
-            HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(handle);
-            ModelAndView modelAndView = handlerAdapter.handle(request, response, handle);
-
+            modelAndView = getModelAndView(request, response);
             View view = modelAndView.getView();
             view.render(modelAndView.getModel(), request, response);
         } catch (Throwable e) {
             LOG.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
+        }
+    }
+
+    private ModelAndView getModelAndView(HttpServletRequest request, HttpServletResponse response)
+        throws Exception {
+        try {
+            Object handle = handlerMappingRegistry.getHandle(request);
+            HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(handle);
+            return handlerAdapter.handle(request, response, handle);
+        } catch (AbstractCustomException e) {
+            return new ModelAndView(new JspView(e.getPages().redirectPageName()));
         }
     }
