+++ mvc/src/main/java/nextstep/mvc/controller/exception/AbstractCustomException.java
@@ -1,0 +1,5 @@
+package nextstep.mvc.controller.exception;
+
+public abstract class AbstractCustomException extends RuntimeException implements CustomException {
+
+}
