+++ mvc/src/main/java/nextstep/mvc/controller/exception/BadRequestException.java
@@ -1,5 +1,11 @@
 package nextstep.mvc.controller.exception;
 
-public class BadRequestException extends RuntimeException{
+import nextstep.mvc.Pages;
 
+public class BadRequestException extends AbstractCustomException {
+
+    @Override
+    public Pages getPages() {
+        return Pages.BAD_REQUEST;
+    }
 }
