+++ mvc/src/main/java/nextstep/mvc/controller/exception/CustomException.java
@@ -1,0 +1,8 @@
+package nextstep.mvc.controller.exception;
+
+import nextstep.mvc.Pages;
+
+public interface CustomException {
+
+    Pages getPages();
+}
