--- mvc/src/main/java/nextstep/mvc/controller/exception/HandleKeyDuplicationException.java
@@ -1,8 +1,0 @@
-package nextstep.mvc.controller.exception;
-
-public class HandleKeyDuplicationException extends RuntimeException {
-
-    public HandleKeyDuplicationException() {
-        super("HandleKeyDuplicate Error!!");
-    }
-}
