+++ mvc/src/main/java/nextstep/mvc/controller/exception/InternalServerException.java
@@ -1,5 +1,11 @@
 package nextstep.mvc.controller.exception;
 
-public class InternalServerException extends RuntimeException{
+import nextstep.mvc.Pages;
 
+public class InternalServerException extends AbstractCustomException {
+
+    @Override
+    public Pages getPages() {
+        return Pages.INTERNAL_SERVER;
+    }
 }
