--- mvc/src/main/java/nextstep/mvc/controller/exception/UnauthorizedException.java
@@ -1,5 +1,0 @@
-package nextstep.mvc.controller.exception;
-
-public class UnauthorizedException extends RuntimeException {
-
-}
