+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecution.java
@@ -4,12 +4,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.lang.reflect.Method;
 import nextstep.mvc.view.ModelAndView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class HandlerExecution {
-
-    private static final Logger LOG = LoggerFactory.getLogger(HandlerExecution.class);
 
     private final Object instance;
     private final Method method;
