+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerMappingRegistry.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.controller.exception.BadRequestException;
 
 public class HandlerMappingRegistry {
 
@@ -27,6 +28,6 @@ public class HandlerMappingRegistry {
             .map(handlerMapping -> handlerMapping.getHandler(request))
             .filter(Objects::nonNull)
             .findFirst()
-            .orElseThrow();
+            .orElseThrow(BadRequestException::new);
     }
 }
